+++ src/Configuration.Writable.Abstractions/IReadOnlyOptions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Microsoft.Extensions.Options;
 
 namespace Configuration.Writable;
 
@@ -13,19 +12,11 @@ public interface IReadOnlyOptions<T> : IReadOnlyOptionsCore<T>
 {
     /// <summary>
     /// Returns the current <typeparamref name="T"/> instance.
-    /// This method behaves similarly to the <see cref="IOptionsMonitor{T}.CurrentValue"/> method.
     /// </summary>
     T CurrentValue { get; }
 
     /// <summary>
-    /// Retrieves the configuration settings object for the default configuration section.
-    /// </summary>
-    WritableOptionsConfiguration<T> GetOptionsConfiguration();
-
-    /// <summary>
-    /// Registers a listener to be called whenever a named <typeparamref name="T"/> changes. <br/>
-    /// This method behaves similarly to the <see cref="IOptionsMonitor{T}.OnChange"/> method, <br/>
-    /// but is only called for the default instanceName.
+    /// Registers a listener to be called whenever the default <typeparamref name="T"/> changes.
     /// </summary>
     /// <param name="listener">The action to be invoked when <typeparamref name="T"/> has changed.</param>
     /// <returns>An <see cref="IDisposable"/> which should be disposed to stop listening for changes.</returns>