+++ src/Configuration.Writable.Abstractions/IReadOnlyOptionsMonitor.cs
@@ -1,5 +1,4 @@
 using System;
-using Microsoft.Extensions.Options;
 
 namespace Configuration.Writable;
 
@@ -7,16 +6,13 @@ namespace Configuration.Writable;
 /// Represents a read-only configuration options monitor interface for accessing and monitoring options of type <typeparamref name="T"/>.
 /// </summary>
 /// <typeparam name="T">The type of the options class.</typeparam>
-public interface IReadOnlyOptionsMonitor<T>
-    : IReadOnlyOptions<T>,
-        IReadOnlyNamedOptions<T>,
-        IOptionsMonitor<T>
+public interface IReadOnlyOptionsMonitor<T> : IReadOnlyOptions<T>, IReadOnlyNamedOptions<T>
     where T : class, new()
 {
     /// <inheritdoc cref="IReadOnlyOptions{T}.CurrentValue" />
     new T CurrentValue { get; }
 
-    /// <inheritdoc cref="IReadOnlyNamedOptions{T}.Get(string?)" />
+    /// <inheritdoc cref="IReadOnlyNamedOptions{T}.Get(string)" />
     new T Get(string name);
 
     /// <inheritdoc cref="IReadOnlyOptionsCore{T}.OnChange" />