+++ src/Configuration.Writable.Core/WritableOptionsCoreExtensions.cs
@@ -68,6 +68,18 @@ public static class WritableOptionsCoreExtensions
         services.AddSingleton<IWritableOptionsMonitor<T>>(p =>
             p.GetRequiredService<WritableOptionsImpl<T>>()
         );
+        services.AddSingleton<IOptionsConfigurationAccessor<T>>(p =>
+            p.GetRequiredService<WritableOptionsImpl<T>>()
+        );
+        services.AddSingleton<INamedOptionsConfigurationAccessor<T>>(p =>
+            p.GetRequiredService<WritableOptionsImpl<T>>()
+        );
+        services.AddSingleton<IWritableOptionsConfigurationAccessor<T>>(p =>
+            p.GetRequiredService<WritableOptionsImpl<T>>()
+        );
+        services.AddSingleton<INamedWritableOptionsConfigurationAccessor<T>>(p =>
+            p.GetRequiredService<WritableOptionsImpl<T>>()
+        );
 
         // if named instance, add named wrapper
         if (!string.IsNullOrEmpty(instanceName))
@@ -83,6 +95,18 @@ public static class WritableOptionsCoreExtensions
             services.AddKeyedSingleton<IReadOnlyOptions<T>>(
                 instanceName,
                 (p, key) => p.GetRequiredKeyedService<IWritableOptions<T>>(instanceName)
+            );
+            services.AddKeyedSingleton<IOptionsConfigurationAccessor<T>>(
+                instanceName,
+                (p, key) =>
+                    (IOptionsConfigurationAccessor<T>)
+                        p.GetRequiredKeyedService<IWritableOptions<T>>(instanceName)
+            );
+            services.AddKeyedSingleton<IWritableOptionsConfigurationAccessor<T>>(
+                instanceName,
+                (p, key) =>
+                    (IWritableOptionsConfigurationAccessor<T>)
+                        p.GetRequiredKeyedService<IWritableOptions<T>>(instanceName)
             );
             services.AddKeyedSingleton<IReadOnlyOptionsMonitor<T>>(
                 instanceName,